% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderlyweb.R
\name{orderlyweb}
\alias{orderlyweb}
\title{Create an orderlyweb object}
\usage{
orderlyweb(..., api_client = NULL)
}
\arguments{
\item{...}{Arguments passed through to
\code{\link{orderlyweb_api_client}}}

\item{api_client}{A pre-constructed
\code{\link{orderlyweb_api_client}}, in which case \code{...} is
ignored}
}
\description{
Create an orderlyweb object, for interaction with the OrderlyWeb
API at a high level.  For low level access, see
\code{\link{orderlyweb_api_client}} or the \code{$api_client}
field of the created object.
}
\examples{
cl <- orderlyweb::orderlyweb(host = "example.com", port = 443,
                             token = "mytoken")
cl$api_client$is_authorised()
}
