% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{orderlyweb_api_client}
\alias{orderlyweb_api_client}
\title{Create a low-level OrderlyWeb client}
\usage{
orderlyweb_api_client(hostname, port, token, https = TRUE,
  prefix = NULL, api_version = 1, insecure = FALSE,
  verbose = FALSE)
}
\arguments{
\item{hostname}{Fully qualified hostname for the OrderlyWeb instance}

\item{port}{Port to use}

\item{token}{Your application token for authentication}

\item{https}{Optional logical, indicating if this is an https
connection - this should be \code{TRUE} in all production
settings or credentials will be sent in the clear!}

\item{prefix}{A prefix, if your OrderlyWeb server is mounted at a
path within some larger website.}

\item{api_version}{The API version to request - this should be
left as 1.}

\item{insecure}{Avoid SSL certificate testing - this is completely
insecure (as bad as http) and exists only for testing.}

\item{verbose}{Be verbose in all http exchanges.  This will be
very noisy.}
}
\description{
Create a low-level HTTP API client for use with OrderlyWeb.
}
\examples{
cl <- orderlyweb::orderlyweb_api_client(host = "example.com", port = 443,
                                        token = "mytoken")
cl$is_authorised()
}
