% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly.R
\name{orderlyweb_remote}
\alias{orderlyweb_remote}
\title{Create orderly remote}
\usage{
orderlyweb_remote(hostname, port, token, https = TRUE, prefix = NULL)
}
\arguments{
\item{hostname}{Fully qualified hostname for the OrderlyWeb instance}

\item{port}{Port to use}

\item{token}{Your application token for authentication}

\item{https}{Optional logical, indicating if this is an https
connection - this should be \code{TRUE} in all production
settings or credentials will be sent in the clear!}

\item{prefix}{A prefix, if your OrderlyWeb server is mounted at a
path within some larger website.}
}
\description{
Implements an orderly "remote" using OrderlyWeb as a backend.  Use
this within an \code{orderly_config.yml} configuration.
}
\examples{
remote <- orderlyweb::orderlyweb_remote("example.com", 443, "mytoken")
remote
}
